==> 5-Association Testing.R <==
#** -------- Association Testing -------- **#
# ========================================= #
#    Zach Ribau  |  June 27, 2024       

#**_____________________________________________________**#
## -------------------- CONFIG ------------------------- ##

source(file.path(dirname(rstudioapi::getActiveDocumentContext()$path), "Config", "config.R"))

Output_Dir <- NA

#**__________________________________________________________**#
## ---------- 1. ADULT - SNP ASSOCIATION MODELS ------------- ##


## ----------- 1.1 Env. Variables and Model Functions  ------------- ##

#Adult datasert
adult_data <- Database[["Complete Datasets"]][["Adults"]]
children_data <- Database[["Complete Datasets"]][["Children"]]
full_cohort_data <- Database[["Complete Datasets"]][["Full Cohort"]]

#Functions to automate the model fitting and summarizing process

#FOR GENERAL LINEAR MODELS USED IN SEPARATE COHORTS
fit_and_summarize_model <- function(response, snp_columns, covariates, data) {
  # Create the formula for the model
  snp_formula <- paste(snp_columns, collapse = " + ")
  response_formula <- paste(response, "~")
  covariates_str <- paste(covariates, collapse = " + ")
  full_formula <- as.formula(paste(response_formula, snp_formula, "+", covariates_str))
  
  #Filter the data for missing data in our SNP columns
  filtered_data <- data %>%
    filter(if_all(all_of(c(snp_columns, covariates)), ~ !is.na(.) & !is.nan(.) & !is.infinite(.)))
  
  # Fit the model
  model <- glm(
    formula = full_formula,
    data = filtered_data,
    family = "gaussian"
  )
  
  # Return the summary of the model
  return(summary(model))
}

#FOR GENERAL ESTIMATING EQUATIONS USED IN FULL COHORT
fit_and_summarize_gee_model <- function(response, snp_columns, covariates, data, id_column, correlation_structure = "exchangeable") {
  # Create the formula for the model
  snp_formula <- paste(snp_columns, collapse = " + ")
  response_formula <- paste(response, "~")
  covariates_str <- paste(covariates, collapse = " + ")
  full_formula <- as.formula(paste(response_formula, snp_formula, "+", covariates_str))
  
  # Filter the data for missing data in our SNP columns
  filtered_data <- data %>%
    filter(if_all(all_of(c(snp_columns, covariates, response)), ~ !is.na(.) & !is.nan(.) & !is.infinite(.)))
  
  # Fit the GEE model
  gee_model <- geeglm(
    formula = full_formula,
    data = filtered_data,
    id = filtered_data[[id_column]],  # Specify the cluster variable
    family = gaussian(),  # Specify the family; adjust as needed
    corstr = correlation_structure  # Specify the correlation structure
  )
  
  # Return the summary of the model
  return(summary(gee_model))
}

#Function that extracts significant SNP ids from model summary objects within lists generated by functions above
extract_significant_snps <- function(model_summaries, snp_columns, p_value_threshold = p_value_threshold) {
  significant_snps_list <- list()
  
  # Loop over the named models and extract significant SNPs for diet
  for (response in names(model_summaries)) {
    model_summary <- model_summaries[[response]]
    
    if (is.null(model_summary)) {
      warning("Model summary for response:", response, "is NULL")
      next
    }
    
    # Extract coefficients and p-values
    coefficients <- model_summary$coefficients
    
    # Determine the correct p-value column
    p_value_column <- if ("Pr(>|W|)" %in% colnames(coefficients)) {
      "Pr(>|W|)"
    } else if ("Pr(>|t|)" %in% colnames(coefficients)) {
      "Pr(>|t|)"
    } else {
      warning("P-values column not found in the coefficients for response:", response)
      next
    }
    
    # Ensure that snp_columns exist in the row names of coefficients
    existing_snp_columns <- intersect(snp_columns, rownames(coefficients))
    
    if (length(existing_snp_columns) > 0) {
      snp_p_values <- coefficients[existing_snp_columns, p_value_column, drop = FALSE]
      
      # Find significant SNPs
      sig_snps <- rownames(snp_p_values)[snp_p_values < p_value_threshold]
      
      # Add to the list of significant SNPs
      significant_snps_list[[response]] <- sig_snps
    } else {
      warning("None of the SNP columns are present in the coefficients for response:", response)
    }
  }
  
  # Flatten the list and remove duplicates for diet
  unique_significant_snps <- unique(unlist(significant_snps_list))
  return(unique_significant_snps)
}

#Function to extract significant SNPs for single phenotype
get_significant_snps_trait <- function(model_summary, snp_columns, p_value_threshold = 0.05) {
  # Extract coefficients and p-values
  coefficients <- model_summary$coefficients
  
  # Determine the correct p-value column
  p_value_column <- if ("Pr(>|W|)" %in% colnames(coefficients)) {
    "Pr(>|W|)"
  } else {
    "Pr(>|t|)"
  }
  
  # Ensure that snp_columns exist in the row names of coefficients
  existing_snp_columns <- intersect(snp_columns, rownames(coefficients))
  snp_p_values <- coefficients[existing_snp_columns, p_value_column, drop = FALSE]
  
  # Find significant SNPs
  sig_snps <- rownames(snp_p_values)[snp_p_values < p_value_threshold]
  
  # Return the list of significant SNPs
  return(sig_snps)
}


## ----------------- 1.2 Diet Phenotype Models ----------------- ##

diet_models_adult <- setNames(lapply(diet_response_columns, function(response) {
  fit_and_summarize_model(response, snp_columns, covariates, adult_data)
}), diet_response_columns)

diet_sig_snps_adult <- extract_significant_snps(diet_models_adult, snp_columns, p_value_threshold = p_value_threshold)

#Extract detailed information here!


## ----------------- 1.3 Sleep Phenotype Models ----------------- ##

# Fit models for sleep and name them by response variable
sleep_models_adult <- setNames(lapply(sleep_response_columns, function(response) {
  fit_and_summarize_model(response, snp_columns, covariates, adult_data)
}), sleep_response_columns)

sleep_sig_snps_adult <- extract_significant_snps(sleep_models_adult, snp_columns, p_value_threshold = p_value_threshold)

#Extract detailed information here!


## ----------------- 1.4 All significant SNPs ----------------- ##

#Prepare organized dataframe from list of SNPs
all_significant_snps_adult <- unique(c(diet_sig_snps_adult, sleep_sig_snps_adult))
adult_sig_snps_df <- data.frame(SNP = all_significant_snps_adult, stringsAsFactors = FALSE) %>%
  mutate(
    diet = SNP %in% diet_sig_snps_adult,
    sleep = SNP %in% sleep_sig_snps_adult
  ) %>%
  mutate(
    Phenotype = case_when(
      diet & sleep ~ "Both",
      sleep ~ "Sleep",
      diet ~ "Diet",
      TRUE ~ "None"
    )
  ) %>%
  select(-diet, -sleep)

#Loop over diet response columns and find sig snps within each
adult_sig_snps_diet_trait <- list()

#Loop over each diet column to apply the function
for (col in diet_response_columns) {
  adult_sig_snps_diet_trait[[col]] <- get_significant_snps_trait(diet_models_adult[[col]], snp_columns, p_value_threshold = p_value_threshold)
}

#Loop over diet response columns and find sig snps within each
adult_sig_snps_sleep_trait <- list()

#Loop over each diet column to apply the function
for (col in sleep_response_columns) {
  adult_sig_snps_sleep_trait[[col]] <- get_significant_snps_trait(sleep_models_adult[[col]], snp_columns, p_value_threshold = p_value_threshold)
}

#**__________________________________________________________**#
## ---------- 2. CHILD - SNP ASSOCIATION MODELS ------------- ##

## ----------------- 2.1 Diet Phenotype Models ----------------- ##

diet_models_children <- setNames(lapply(diet_response_columns, function(response) {
  fit_and_summarize_model(response, snp_columns, covariates, children_data)
}), diet_response_columns)

diet_sig_snps_children <- extract_significant_snps(diet_models_children, snp_columns, p_value_threshold = p_value_threshold)

#Extract detailed information here!


## ----------------- 2.2 Sleep Phenotype Models ----------------- ##

# Fit models for sleep and name them by response variable
sleep_models_children <- setNames(lapply(sleep_response_columns, function(response) {
  fit_and_summarize_model(response, snp_columns, covariates, children_data)
}), sleep_response_columns)

sleep_sig_snps_children <- extract_significant_snps(sleep_models_children, snp_columns, p_value_threshold = p_value_threshold)

#Extract detailed information here!

## ----------------- 2.3 All significant SNPs ----------------- ##

#Prepare organized dataframe from list of SNPs
all_significant_snps_children <- unique(c(diet_sig_snps_children, sleep_sig_snps_children))
children_sig_snps_df <- data.frame(SNP = all_significant_snps_children, stringsAsFactors = FALSE) %>%
  mutate(
    diet = SNP %in% diet_sig_snps_children,
    sleep = SNP %in% sleep_sig_snps_children
  ) %>%
  mutate(
    Phenotype = case_when(
      diet & sleep ~ "Both",
      sleep ~ "Sleep",
      diet ~ "Diet",
      TRUE ~ "None"
    )
  ) %>%
  select(-diet, -sleep)

#Loop over diet response columns and find sig snps within each
child_sig_snps_diet_trait <- list()

#Loop over each diet column to apply the function
for (col in diet_response_columns) {
  child_sig_snps_diet_trait[[col]] <- get_significant_snps_trait(diet_models_children[[col]], snp_columns, p_value_threshold = p_value_threshold)
}

#Loop over diet response columns and find sig snps within each
child_sig_snps_sleep_trait <- list()

#Loop over each diet column to apply the function
for (col in sleep_response_columns) {
  child_sig_snps_sleep_trait[[col]] <- get_significant_snps_trait(sleep_models_children[[col]], snp_columns, p_value_threshold = p_value_threshold)
}

#**______________________________________________________________**#
## ---------- 3. FULL COHORT SNP ASSOCIATION MODELS ------------- ##

## ----------------- 3.1 Diet Phenotype Models ----------------- ##

diet_models_full <- setNames(lapply(diet_response_columns, function(response) {
  fit_and_summarize_gee_model(response, snp_columns, covariates, full_cohort_data, id_column = "fid")
}), diet_response_columns)

full_diet_sig_snps <- extract_significant_snps(diet_models_full, snp_columns, p_value_threshold = p_value_threshold)

## ----------------- 3.2 Sleep Phenotype Models ----------------- ##

sleep_models_full <- setNames(lapply(sleep_response_columns, function(response) {
  fit_and_summarize_gee_model(response, snp_columns, covariates, full_cohort_data, id_column = "fid")
}), sleep_response_columns)

full_sleep_sig_snps <- extract_significant_snps(sleep_models_full, snp_columns, p_value_threshold = p_value_threshold)

## ----------------- 3.3 All significant SNPs ----------------- ##

all_significant_snps_full <- unique(c(full_diet_sig_snps, full_sleep_sig_snps))

full_sig_snps_df <- data.frame(SNP = all_significant_snps_full, stringsAsFactors = FALSE) %>%
  mutate(
    diet = SNP %in% full_diet_sig_snps,
    sleep = SNP %in% full_sleep_sig_snps
  ) %>%
  mutate(
    Phenotype = case_when(
      diet & sleep ~ "Both",
      sleep ~ "Sleep",
      diet ~ "Diet",
      TRUE ~ "None"
    )
  ) %>%
  select(-diet, -sleep)

#Loop over diet response columns and find sig snps within each
full_sig_snps_diet_trait <- list()

#Loop over each diet column to apply the function
for (col in diet_response_columns) {
  full_sig_snps_diet_trait[[col]] <- get_significant_snps_trait(diet_models_full[[col]], snp_columns, p_value_threshold = p_value_threshold)
}

#Loop over diet response columns and find sig snps within each
full_sig_snps_sleep_trait <- list()

#Loop over each diet column to apply the function
for (col in sleep_response_columns) {
  full_sig_snps_sleep_trait[[col]] <- get_significant_snps_trait(sleep_models_full[[col]], snp_columns, p_value_threshold = p_value_threshold)
}

'Significant SNPS' <- list(Adult = list(Diet = adult_sig_snps_diet_trait, Sleep = adult_sig_snps_sleep_trait), Children = list(Diet = child_sig_snps_diet_trait, Sleep = child_sig_snps_sleep_trait), 'Full Cohort' = list(Diet = full_sig_snps_diet_trait, Sleep = full_sig_snps_sleep_trait))

#**______________________________________________________________**#
## -------------- 4. ORGANIZING SNP ASSOCIATIONS ---------------- ##

#Function to get significant snp associations and their pvalues from a given model object and trait
get_snp_pvalues <- function(model_name, model_summary, snp_columns, min_p_threshold = 0.05) {
  # Ensure model_summary is a list with a 'coefficients' element
  if (!is.list(model_summary) || !("coefficients" %in% names(model_summary))) {
    stop("model_summary should be a list containing a 'coefficients' element.")
  }
  
  # Filter the coefficients to include only those in snp_columns
  filtered_coefficients <- model_summary$coefficients[rownames(model_summary$coefficients) %in% snp_columns, ]
  
  # Extract the 'Pr(>|W|)' column
  p_values <- filtered_coefficients[, p_col]
  
  # Create a data frame with SNP names and their corresponding p-values
  df <- data.frame(SNP = rownames(filtered_coefficients), PValue = p_values, stringsAsFactors = FALSE)
  
  # Filter the data frame to include only rows with p-values greater than the threshold
  filtered_df <- df[df$PValue <= min_p_threshold, ]
  
  # Return the filtered data frame
  return(filtered_df)
}

#Function to apply above function over series of traits within model object, saving to a dataframe
apply_snp_pvalues <- function(response_columns, model_set, snp_columns, model_name) {
  results_list <- list()
  
  # Loop over each response column to apply the function
  for (response in response_columns) {
    result <- get_snp_pvalues(response, model_set[[response]], snp_columns)
    # Check if the result is not NULL or empty
    if (!is.null(result) && nrow(result) > 0) {
      results_list[[response]] <- result
    }
  }
  
  # Combine the results into a single data frame if there are any results
  if (length(results_list) > 0) {
    full_df <- do.call(rbind, lapply(names(results_list), function(response) {
      df <- results_list[[response]]
      df$Trait <- response
      df$Model <- model_name
      return(df)
    }))
  } else {
    full_df <- data.frame()
  }
  
  return(full_df)
}

#Specify p_col
p_col <- "Pr(>|W|)"

#Apply nested function setup to full cohort diet and sleep associations
diet_full <- apply_snp_pvalues(diet_response_columns, diet_models_full, snp_columns, "diet_models_full")
sleep_full <- apply_snp_pvalues(sleep_response_columns, sleep_models_full, snp_columns, "sleep_models_full")

p_col <- "Pr(>|t|)"

#And to adult and children datasets
diet_adult <- apply_snp_pvalues(diet_response_columns, diet_models_adult, snp_columns, "diet_models_adult")
sleep_adult <- apply_snp_pvalues(sleep_response_columns, sleep_models_adult, snp_columns, "sleep_models_adult")

diet_children <- apply_snp_pvalues(diet_response_columns, diet_models_children, snp_columns, "diet_models_children")
sleep_children <- apply_snp_pvalues(sleep_response_columns, sleep_models_children, snp_columns, "sleep_models_children")

data_frames <- list(diet_full, sleep_full, diet_adult, sleep_adult, diet_children, sleep_children)
snp_associations <- do.call(rbind, data_frames)
rownames(snp_associations) <- NULL

snp_summary <- snp_associations %>%
  group_by(SNP) %>%
  summarise()

snp_associations <- snp_associations %>%
  mutate(
    Model = ifelse(Model %in% c("diet_models_full", "sleep_models_full"), "full cohort",
                   ifelse(Model %in% c("diet_models_adult", "sleep_models_adult"), "adult",
                          ifelse(Model %in% c("diet_models_children", "sleep_models_children"), "children", Model)))
  )
